+++ NEWS.md
@@ -1,11 +1,19 @@
 # Change log of the R package 'via'
 
-# via 0.2.0. (Junky Jackal) 
+# via 0.2.0. 
 ### Added 
-- Class: SfcArray between the inheritance order of XArray -> SfArray, to allow geometry-only arrays
+- Class: `SfcArray` between the inheritance order of `XArray` -> `SfArray`, to allow geometry-only arrays
+- casting functions between `SfcArray`, `SfArray`, and `XArray`
+
+### Changed
+- The `paleocoastlines` demo data is now an `SfcArray`-class object. Column names changed to `c("margin", "coast")`. 
+
+### Fixed
+- alternative generic of `st_bbox()` changed to be a function of `obj` instead of `x` for generic-method consistency.
+- removed deprecated documentation reference to 'raster'
 
 
-# via 0.1.0. (Initial Ibex) - 2023-03-15
+# via 0.1.0. - 2023-03-15
 ### Added 
 - Classes: XArray, SfArray, RasterArray
 - Suggests: terra, sf